requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.11
django-environ==0.9.0
djangorestframework==3.13.1
httplib2==0.20.4
idna==3.3
pyparsing==3.0.8
python-environ==0.4.54
pytz==2022.1
requests==2.27.1
slugify==0.0.1
sqlparse==0.4.2
urllib3==1.26.9
xmltodict==0.13.0
